full draft (...again)
this commit has my redone first full draft of my webpages. it also has my edited .xcf file where I edited the logo of my "company" on GIMP to be shorter. I have changed all of the HTML files to restructure the nav bar, redo the photo grid on the index page, add a table for the admission page, add an embedded Google Map in contact, update CSS like removing the background image, and adjust my assets page for the sources that I have used. I've compressed this file in another location on my laptop just in case it goes wrong again, but I'm glad I was able to recreate it in a few hours after losing a week's progress this afternoon!